Resolved Json/save method, added Instance class
The saving page has been updated to be fully-functional, converting the Instance class into a local Json file. An Instance class was added to house the attributes, classes, and relationships arrays, but will likely need further modification as these classes are further elaborated upon.
